OSIS-4616 move tree persist prerequisites to own module
