+++ program_management/tests/ddd/repositories/test_persist_prerequisite.py
@@ -21,6 +21,8 @@
 #  at the root of the source code of this program.  If not,
 #  see http://www.gnu.org/licenses/.
 # ############################################################################
+from unittest import mock
+
 from django.test import TestCase
 
 from base.models.prerequisite import Prerequisite
@@ -31,8 +33,24 @@ from base.tests.factories.learning_unit_year import LearningUnitYearFactory
 from base.tests.factories.prerequisite import PrerequisiteFactory
 from program_management.ddd.domain import prerequisite
 from program_management.ddd.domain.prerequisite import NullPrerequisite
-from program_management.ddd.repositories import persist_prerequisite
+from program_management.ddd.repositories import _persist_prerequisite
+from program_management.tests.ddd.factories.link import LinkFactory
 from program_management.tests.ddd.factories.node import NodeEducationGroupYearFactory, NodeLearningUnitYearFactory
+from program_management.tests.ddd.factories.program_tree import ProgramTreeFactory
+
+
+class TestPersist(TestCase):
+    @mock.patch("program_management.ddd.repositories._persist_prerequisite._persist")
+    def test_call_persist_prerequisite_on_changed_node(self, mock_persist_prerequisite):
+        tree = ProgramTreeFactory()
+        link1 = LinkFactory(parent=tree.root_node, child=NodeLearningUnitYearFactory())
+        link2 = LinkFactory(parent=tree.root_node, child=NodeLearningUnitYearFactory())
+
+        link1.child.set_prerequisite(NullPrerequisite())
+
+        _persist_prerequisite.persist(tree)
+
+        mock_persist_prerequisite.assert_called_once_with(tree.root_node, link1.child)
 
 
 class TestPersistPrerequisite(TestCase):
@@ -49,7 +67,7 @@ class TestPersistPrerequisite(TestCase):
 
     def test_when_null_prerequisite_given(self):
         self.node.prerequisite = NullPrerequisite()
-        persist_prerequisite.persist(self.root_node, self.node)
+        _persist_prerequisite._persist(self.root_node, self.node)
 
         prerequisite_obj = Prerequisite.objects.get(education_group_year__id=self.root_node.node_id)
         self.assertFalse(
@@ -62,7 +80,7 @@ class TestPersistPrerequisite(TestCase):
             self.current_academic_year.year
         )
         self.node.prerequisite = prerequisite_obj
-        persist_prerequisite.persist(self.root_node, self.node)
+        _persist_prerequisite._persist(self.root_node, self.node)
 
         prerequisite_obj = Prerequisite.objects.get(education_group_year__id=self.root_node.node_id)
         self.assertTrue(
@@ -76,7 +94,7 @@ class TestPersistPrerequisite(TestCase):
             items__groups=((self.luy1,), (self.luy2,))
         )
         self.node.prerequisite = NullPrerequisite()
-        persist_prerequisite.persist(self.root_node, self.node)
+        _persist_prerequisite._persist(self.root_node, self.node)
 
         prerequisite_obj = Prerequisite.objects.get(education_group_year__id=self.root_node.node_id)
         self.assertFalse(
